Initial commit
The main "CtorSet" and "With" extension methods are pretty much done,
with integration with the Optional struct to ensure that null references
are not allowed anywhere that the type system does not expressly dictate
that they may be (exceptions will be thrown at runtime if these rules
are not followed). There is a Set class which is integrated with the
extension methods - its functionality is quite minimal but it's a good
starting point.