Limited "With" methods to IAmImmutable targets
While the "With" methods do not HAVE to be limited to
IAmImmutable-implementing classes (unlike "CtorSet", which should ONLY
be used on classes that are designed for it, so that private setters can
not get called outside of the class), it's probably better not to
pollute the "available methods, properties and extensions methods" list
for ALL types (in places where the ProductiveRage.Immutable namespace
has been imported). I might go back on this at a later date, I'm not
100% sure that being more restrictive is the correct design in this
case.