Hacked up a batch file to build a library+analyser NuGet package
This is probably not the most elegant way to do this, but it will do the
job for now. (It's challenging with Bridge libraries because no single
project can load the Bridge library AND the Analyser, so it's not
straight-forward to have one project with an "AfterBuild" step that
generates the NuGet package).