Downgraded CodeAnalysis NuGet dependency to work with VS2015 pre-Update1
The project type that I added to create the analyser would only work
with VS2015 Update 1 so I've downgraded some dependencies, as advised at
https://github.com/dotnet/roslyn/issues/4831

Note that this seems to have introduced a warning that is talked about
at https://github.com/dotnet/roslyn/issues/7276 (it was fixed in the
later versions of the packages, but these are the versions that require
Update 1). The warning is annoying but seems harmless.