Can't share analyser functionality in an abstract base class
The refactor to move common code into a base class caused warnings when
the NuGet package was installed from the abstract base class as
apparently you "Cannot create an abstract class" (meaning, I think, that
DiagnosticAnalyzer should not be derived by an abstract class). It
didn't break the functionality since the other analysers worked, it was
only when VS attempted to use the base class analyser directly (which
explains why the unit tests and everything worked, it was only when
actually installed as a package). Treats me right for using inheritance
when I didn't absolutely need to.