Private setters are not reported in meta from external assemblies, so they can't be reported as errors
This means that external assemblies could have invalid classes, which is
a problem - but there's no way to be sure from the consuming code. The
best solution may to be add an IAmImmutable analyser which ensures that
all properties within an IAmImmutable implementation follow the property
guidelines. This would prevent any properties from being truly read-only
(they would HAVE to have a setter at all times), so there could not be
computed properties. It would also not prevent runtime errors arising
from assemblies built in an environment where the analysers are not
used.