Started analyser for IAmImmutable implementations
Since it's not possible to verify that all properties with public
getters have setters on types that implement IAmImmutable, it will be
beneficial to have analysers that pick up these problems when the
classes are being written, rather than only when they're being consumed.
(If an IAmImmutable implementation exists in an external library then
any private getters that properties may have will not be reported in the
library's meta data).