Correction to Set class that would affect it if compiled as C# but that doesn't as JavaScript
The negative-counting loop used by Set's Add method has a logic error
that fails in C# (it rolled over when there was a single item to
persist). This doesn't affect the JavaScript version but it's incorrect
and Bridge might change how this works in the future, so it's worth
fixing now.