Added a rule to disallow public setters on IAmImmutable implementations and bumped version numbers for release
Public setters on IAmImmutable implementations are a mistake, most
likely accidental (the class can't be immutable if it has a public
setter, since that means that the data may be changed!).

Bumped Bridge dependency from 1.11.1 to 1.12.1, I'm not 100% sure that
there are any changes that are required but 1.12.1 brought a lot of
fixes that are so beneficial that it makes no sense to support working
with 1.11.1 any more.