Don't warn about Validate calls being missing from IAmImmutable ctor's if Validate doesn't return void
If it returns a non-void type then it is probably not intended to be
used as a throw-if-current-state-is-invalid check that should be called
from each ctor and after every "With" call. Also don't warn if it is a
generic method.. these two edge cases make me suspect that there should
be more logic to "With" calls to ensure that we don't try to call
Validate in either of these case (but I'll worry about that another
day).